build(webpack): rename top-level index.js to candela.js to prevent confusion between import styles
i.e. the following line

````javascript
import candela from 'candela';
````

will import the UMD bundle, while

````javascript
import candela from 'candela/candela';
````

will import the top-level file directly (allowing downstream Webpack to build it
into something probably similar to the original bundle).
